requirements.txt
django == 1.10
python-social-auth
python-twitter